+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.24.0] - 2025-02-09
 ### Added:
 - Docs: Tips for setting custom coordinates for plots.
+- Tree Exporter: `tree_to_pillow_graph` method to export tree to pillow image in graph format.
 ### Changed:
+- Plot: Allow `reverse` argument to allow top-bottom y coordinates in Reingold Tilford algorithm.
 - Docs: Add more elaboration for exporting to image for tree and dag.
 - Misc: Split tree/construct and tree/export into multiple files.
 
@@ -725,7 +729,8 @@ ignore null attribute columns.
 - Utility Iterator: Tree traversal methods.
 - Workflow To Do App: Tree use case with to-do list implementation.
 
-[Unreleased]: https://github.com/kayjan/bigtree/compare/0.23.1...HEAD
+[Unreleased]: https://github.com/kayjan/bigtree/compare/0.24.0...HEAD
+[0.24.0]: https://github.com/kayjan/bigtree/compare/0.23.1...0.24.0
 [0.23.1]: https://github.com/kayjan/bigtree/compare/0.23.0...0.23.1
 [0.23.0]: https://github.com/kayjan/bigtree/compare/0.22.3...0.23.0
 [0.22.3]: https://github.com/kayjan/bigtree/compare/0.22.2...0.22.3